LayoutTests/http/tests/security/xss-DENIED-xsl-document-securityOrigin-expected.txt

CONSOLE MESSAGE: line 32: SecurityError (DOM Exception 18): Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8080". Protocols, domains, and ports must match.
This test passes if it doesn't alert the contents of innocent-victim.html.  
